Fix erroneous test failure
Fixed a single test that was failing due to expecting an empty folder
to exist in the test folder.  Since the folder was empty, it was not
included in the repository.  Added a .gitignore to said folder so that
it will be included in the repository henceforth.

This does not change or "fix" the behavior of autotagical at all, as the
test itself was wrong (when run from a cloned repo).