Refactor and reformat for 1.0.0 release
Reformated code to be closer to PEP8.

Refactored several aspects, mostly in the settings and file_handler
modules.

Eliminated use of the term "categories".  "Tag groups" is used
everywhere to be consistent with TagSpaces usage.

Standardized usage of "condition", "condition set", and "filter".
"condition" is a single logical component, "condition set"
(previously "filter") is a number of conditions linked by /&|, and
a "filter" (previously "filters") is a list of condition sets linked
logically by inclusive or.